Move registration of Open API related bootstrapping
