+++ src/Api/Program.cs
@@ -97,87 +97,7 @@ static void ConfigureWebApplication(WebApplicationBuilder builder, string[] args
     builder.Services.AddProblemDetails();
     builder.Services.AddHealth();
     builder.Services.AddEndpointsApiExplorer();
-    builder.Services.AddSwaggerGen(c =>
-    {
-        c.AddServer(
-            new OpenApiServer
-            {
-                Url = "https://" + (builder.Configuration.GetValue<string>("OpenApi:Host") ?? "localhost"),
-            }
-        );
-        c.AddSecurityDefinition(
-            "Basic",
-            new OpenApiSecurityScheme
-            {
-                Description = "RFC8725 Compliant JWT",
-                In = ParameterLocation.Header,
-                Name = "Authorization",
-                Scheme = "Basic",
-                Type = SecuritySchemeType.Http,
-            }
-        );
-        c.AddSecurityRequirement(
-            new OpenApiSecurityRequirement
-            {
-                {
-                    new OpenApiSecurityScheme
-                    {
-                        Reference = new OpenApiReference { Type = ReferenceType.SecurityScheme, Id = "Basic" },
-                    },
-                    []
-                },
-            }
-        );
-        c.IncludeXmlComments(Assembly.GetExecutingAssembly());
-        c.IncludeXmlComments(typeof(ImportPreNotification).Assembly);
-        c.SchemaFilter<PossibleValueSchemaFilter>();
-        c.SchemaFilter<JsonConverterSchemaFilter>();
-        c.OperationFilter<PossibleValueOperationFilter>();
-
-        var typeMap = new Dictionary<string, string>
-        {
-            // ReSharper disable once RedundantNameQualifier
-            { typeof(Defra.TradeImportsDataApi.Domain.Ipaffs.CommodityCheck).FullName!, "NotificationCommodityCheck" },
-            {
-                typeof(Defra.TradeImportsDataApi.Domain.CustomsDeclaration.CommodityCheck).FullName!,
-                "CustomsCommodityCheck"
-            },
-        };
-        c.CustomSchemaIds(x =>
-        {
-            var schemaId = x.FullName!;
-
-            if (schemaId.StartsWith("Defra"))
-            {
-                var typeName = typeMap.TryGetValue(x.FullName!, out var mappedTypeName) ? mappedTypeName : x.Name;
-                schemaId = "Defra.TradeImportsDataApi." + typeName;
-            }
-
-            return schemaId;
-        });
-
-        c.SupportNonNullableReferenceTypes();
-        c.UseAllOfToExtendReferenceSchemas();
-        c.SwaggerDoc(
-            "v1",
-            new OpenApiInfo
-            {
-                Description = "TBC",
-                Contact = new OpenApiContact
-                {
-                    Email = "tbc@defra.gov.uk",
-                    Name = "DEFRA",
-                    Url = new Uri(
-#pragma warning disable S1075
-                        "https://www.gov.uk/government/organisations/department-for-environment-food-rural-affairs"
-#pragma warning restore S1075
-                    ),
-                },
-                Title = "Trade Imports Data API",
-                Version = "v1",
-            }
-        );
-    });
+    builder.Services.AddOpenApi(builder.Configuration);
     builder.Services.AddHttpClient();
     builder.Services.AddHeaderPropagation(options =>
     {
@@ -218,17 +138,7 @@ static WebApplication BuildWebApplication(WebApplicationBuilder builder)
     app.MapCustomsDeclarationEndpoints();
     app.MapProcessingErrorEndpoints();
     app.MapSearchEndpoints();
-    app.UseSwagger(options =>
-    {
-        options.RouteTemplate = "/.well-known/openapi/{documentName}/openapi.json";
-    });
-    app.UseReDoc(options =>
-    {
-        options.ConfigObject.ExpandResponses = "200";
-        options.DocumentTitle = "Trade Import Data API";
-        options.RoutePrefix = "redoc";
-        options.SpecUrl = "/.well-known/openapi/v1/openapi.json";
-    });
+    app.UseOpenApi();
     app.UseStatusCodePages();
     app.UseExceptionHandler(
         new ExceptionHandlerOptions
